machine: support wioterminal boards on play.tinygo.org
